web/requirements.txt
Flask
redis
psycopg2-binary
docker